Move helper functions to internal module
